Make project build on Linux
