+++ Sources/Pathos/POSIX/Metadata+Glibc.swift
@@ -1,0 +1,4 @@
+#if canImport(Glibc)
+import Glibc
+
+#endif // canImport(Glibc)
